requirements.txt
python-telegram-bot==20.8
asyncio==3.4.3
requests